requirements.txt
dotenv
langchain_community
langchain_openai
langgraph

pdfplumber
google-cloud-vision



langchain==1.2.10
langchain-text-splitters>=0.0.11

faiss-cpu
